requirements.txt
numpy
torch
Pillow
torch
torchvision
scipy
tensorboardX
opencv-python
matplotlib